requirements.txt
Django>=5.2.8
PyMuPDF>=1.23.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
pytesseract>=0.3.10
easyocr>=1.7.0